Defend against undef issues that occur when version extraction goes awry.
This is performed by using the literal string 'unknown' everywhere
version extraction reports 'undef'.

This choice of wording is because 'undef' is itself a possible version
in a dependency, so reporting that is misleading.

Instead, using 'unknown' to convey "Things broke so bad we cant even
tell if it was undef".